+++ src/components/BurgerConstructor/ConstructorPayment/ConstructorPayment.jsx
@@ -1,6 +1,7 @@
 import styles from "./ConstructorPayment.module.scss";
 import { Button } from "@ya.praktikum/react-developer-burger-ui-components";
 import { CurrencyIcon } from "@ya.praktikum/react-developer-burger-ui-components";
+import PropTypes from "prop-types";
 
 const ConstructorPayment = ({ totalPrice, setOrderVisibility }) => {
   return (
@@ -18,6 +19,11 @@ const ConstructorPayment = ({ totalPrice, setOrderVisibility }) => {
       </Button>
     </div>
   );
+};
+
+ConstructorPayment.propTypes = {
+  totalPrice: PropTypes.number.isRequired,
+  setOrderVisibility: PropTypes.func.isRequired,
 };
 
 export default ConstructorPayment;
