+++ src/index.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.scss";
-import App from "./App";
+import App from "./components/App/App";
 import { Provider } from "react-redux";
 import store from "./services/store";
